avance ejercicio 7 dlp

+++ DLP/Semana4/Guia4/Ejercicio3.py
@@ -1,4 +1,11 @@
 print("Celsius a Fahrenheit")
-c = int(input("Ingrese la temperatura en Celsius: "))
+
+# El usario ingresara aca los datos de los grados Celsius que desea convertir a Fahrenheit.
+c = float(input("Ingrese la temperatura en Celsius: "))
+
+# Formula para convertir de Celsius a Fahrenheit es: F= c * 9/5 +32
+# En este caso utilice 9/5 para multiplicarlo con la variable "c" pero tambien es valido usar 1.8
 f = c * 9/5 + 32
+
+# Esto es lo que se utilizara para mostrar el resultado
 print("la temperatura en fahrenheit es: " + str(f))
